+++ components/Footer.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useRouter } from 'next/router';
-import getLabel from "@/pages/function/GetLabel";
-import { Labels } from "@/pages/function/Types";
+import getLabel from "@/function/GetLabel";
+import { Labels } from "@/function/Types";
 import Link from 'next/link';
 import Image from 'next/image';
 import bilibiliIcon from '@/public/img/bilibili.svg';
